requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
bigcommerce==0.22.2
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pydantic==1.8.2
PyJWT==2.1.0
requests==2.26.0
SQLAlchemy==1.4.23
sqlmodel==0.0.4
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
